% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_manifold.R
\name{add_colorbar3d}
\alias{add_colorbar3d}
\title{Create colorbar for 3D plot}
\usage{
add_colorbar3d(
  center,
  length = 1,
  width = 0.05,
  theta = 0,
  phi = 0,
  zlim = c(0, 1),
  ncolors = 100,
  colfun = colorRampPalette(c("blue", "red")),
  nticks,
  lasttick = TRUE,
  title = NULL,
  cex = 1
)
}
\arguments{
\item{center}{Center position of the colorbar}

\item{length}{Size of color rectangles}

\item{width}{Width of the colorbar}

\item{theta}{Angle for visualisation}

\item{phi}{Angle for visualisation}

\item{zlim}{Limits of the colorbar}

\item{ncolors}{Number of colors in the colorbar}

\item{colfun}{Color function to generate colors}

\item{nticks}{Number of ticks on the colorbar}

\item{lasttick}{If TRUE, add the last tick on the colorbar}

\item{title}{Title of the colorbar}

\item{cex}{Character expansion for text}
}
\value{
A colorbar added to the current RGL 3D plot
}
\description{
Create colorbar for 3D plot
}
