% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{area_choice}
\alias{area_choice}
\title{Minimum Area for stopping criterion by rule of thumb}
\usage{
area_choice(X)
}
\arguments{
\item{X}{The observed data point pattern,
as a \code{\link[spatstat.geom]{ppp.object}}.}
}
\value{
A number.
}
\description{
Minimum Area for stopping criterion by rule of thumb
}
\examples{
area_choice(spatstat.random::rpoispp(100))
}
