% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{as.im.spforest}
\alias{as.im.spforest}
\title{Convert to Pixel Image}
\usage{
\method{as.im}{spforest}(X, ...)
}
\arguments{
\item{X}{A spforest object}

\item{...}{ignored}
}
\value{
A pixel image \code{\link[spatstat.geom]{im.object}}.
}
\description{
Wrapper function to extract imforest from an \code{\link{spforest.object}}.
}
\examples{
forest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
as.im(forest)
}
