% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beisoil}
\alias{beisoil}
\title{Raw soil data for bei}
\format{
A list of 15 \code{im} with different resolutions, and given in the
same order as in Jeff's paper
\describe{
\item{Al}{Al concentration ...}
}
}
\usage{
beisoil
}
\description{
Contains the raw soil data of bei
}
\keyword{datasets}
