% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beisoilnorm}
\alias{beisoilnorm}
\title{Normalised soil data for bei}
\format{
A list of 15 \code{im} with same resolutions, in the
same order as in Jeff's paper, and center and normalised
\describe{
\item{Al}{Al concentration ...}
}
}
\usage{
beisoilnorm
}
\description{
Contains the normalised soil data of bei
}
\keyword{datasets}
