% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beisoilres}
\alias{beisoilres}
\title{soil data for bei}
\format{
A list of 15 \code{im} with same resolutions, and given in the
same order as in Jeff's paper
\describe{
\item{Al}{Al concentration ...}
}
}
\usage{
beisoilres
}
\description{
Contains the soil data of bei put at the same resolution
It has been obtained with the following code
}
\keyword{datasets}
