% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beisoilsmall}
\alias{beisoilsmall}
\title{Downscaled soil data for bei}
\format{
A list of 15 \code{im} with same resolutions, and given in the
same order as in Jeff's paper
\describe{
\item{Al}{Al concentration ...}
}
}
\usage{
beisoilsmall
}
\description{
Contains a downscaled version of the data for testing and vignette
}
\keyword{datasets}
