% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{boxplot.spforest}
\alias{boxplot.spforest}
\title{Boxplot forest}
\usage{
\method{boxplot}{spforest}(x, ...)
}
\arguments{
\item{x}{A \code{\link{spforest.object}} with \code{listcov} not \code{NULL}.}

\item{...}{Ignored.}
}
\value{
Boxplot of the variable importances.
}
\description{
Plot the boxplot of the importance of each covariate
in each tree of the random forest intensity
}
\examples{
forest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
boxplot(forest)
}
