% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{boxplot_spforest}
\alias{boxplot_spforest}
\title{Boxplot forest}
\usage{
boxplot_spforest(x, cores = 1, ...)
}
\arguments{
\item{x}{the forest}

\item{cores}{To compute faster}

\item{...}{why note}
}
\description{
Boxplot forest
}
