% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{extractforest}
\alias{extractforest}
\title{Extract a smaller forest}
\usage{
extractforest(forest, whichtrees)
}
\arguments{
\item{forest}{An object of class \code{\link{spforest}}.}

\item{whichtrees}{A vector indicating which trees to use in the
new forest}
}
\value{
An object of class \code{\link{spforest}}.
}
\description{
Extract a smaller forest
}
\examples{
bigforest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
smallforest <- extractforest(
  forest = bigforest,
  whichtrees = c(2, 5)
)
}
