% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamma_choice}
\alias{gamma_choice}
\title{Tesselation intensityy by rule of thumb}
\usage{
gamma_choice(X)
}
\arguments{
\item{X}{The observed data point pattern,
as a \code{\link[spatstat.geom]{ppp.object}}.}
}
\value{
A number.
}
\description{
Tesselation intensityy by rule of thumb
}
\examples{
gamma_choice(spatstat.random::rpoispp(100))
}
