% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imfcttoreal}
\alias{imfcttoreal}
\title{Convert factor image to real image}
\usage{
imfcttoreal(Z, X)
}
\arguments{
\item{Z}{A pixel image (\code{\link{im.object}}) of type "factor".}

\item{X}{The point pattern from which to compute intensity values.}
}
\value{
A real-valued pixel image. If the image is not factor-valued,
meaning that Z$type is not "factor", then the same image is returned.
}
\description{
Convert a factor-valued image (\code{\link{im.object}}) to a
real-valued image by replacing each factor level
with the corresponding intensity value computed from the point pattern X.
}
\examples{
Z <- beisoilres[[1]]
X <- bei
Y <- cut(Z, quantile(Z), include.lowest = T)
plot(Y)
imfcttoreal(Z = Y, X = X)
}
