% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{importance}
\alias{importance}
\title{Importance of one covariate}
\usage{
importance(forest, id_cov, viptype = 4)
}
\arguments{
\item{forest}{A spforest object}

\item{id_cov}{The index in the list of covariates used in the argument
\code{listcov} passed in \code{\link[spforest]{spforest}}.}

\item{viptype}{An integer in \{1,2,3,4\} representing the type of vip used.}
}
\value{
A vector of the importance of the covariate \code{id_cov}
in the list of covariates of \code{forest}, for each tree.
}
\description{
Importance of one covariate
}
\examples{
forest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
importance(forest, id_cov = 1)
}
