% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{imspforest}
\alias{imspforest}
\title{Compute image of a spatial intensity forest (Deprecated)}
\usage{
imspforest(x, ...)
}
\arguments{
\item{x}{A spatial intensity tree return by RforestPP function}

\item{...}{additional arguments}
}
\value{
A pixel image \code{\link[spatstat.geom]{im}}.
}
\description{
Compute image of a spatial intensity forest (Deprecated)
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
imspforest(forest)
}
