% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_future_seed}
\alias{is_valid_future_seed}
\title{Validate argument compatible with future.seed of future}
\usage{
is_valid_future_seed(seed)
}
\arguments{
\item{seed}{An object to validate.}
}
\value{
A single logical value, \code{TRUE} if \code{seed} is a valid for future.seed,
otherwise \code{FALSE}.
}
\description{
Checks whether an object is a valid value for the \code{future.seed} argument used by
the future framework.
}
\details{
A valid seed is one of:
\itemize{
\item \code{NULL}
\item a single non-missing logical value
\item a single integer value
\item a length-7 integer vector representing an L'Ecuyer-CMRG seed, where
the first element satisfies \code{seed[1] \%\% 10000L == 407L}
}
}
\examples{
is_valid_future_seed(NULL)
is_valid_future_seed(TRUE)
is_valid_future_seed(FALSE)
is_valid_future_seed(c(FALSE, TRUE, TRUE))
is_valid_future_seed(42L)
is_valid_future_seed(c(10407L, 1L, 2L, 3L, 4L, 5L, 6L))
is_valid_future_seed(c(10407L, 1L, 2L, 3L, 4L, 5L, 6L))
is_valid_future_seed(c(10408L, 1L, 2L, 3L, 4L, 5L, 6L))
is_valid_future_seed(c(1L, 2L))
is_valid_future_seed(c(1, 2))
}
