% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_manifold.R
\name{manifold_tree}
\alias{manifold_tree}
\title{Title}
\usage{
manifold_tree(
  intensity,
  vertices,
  faces,
  triangle_centers,
  triangle_areas,
  pointsech
)
}
\arguments{
\item{pointsech}{}
}
\description{
Title
}
