% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.spforest.R
\name{Math.spforestmesh}
\alias{Math.spforestmesh}
\title{Dispatch standard math functions}
\usage{
\method{Math}{spforestmesh}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spforestmesh}}

\item{...}{Additional arguments passed to the function}
}
\value{
An object of class \code{spforestmesh}
}
\description{
Dispatch standard math functions
}
\examples{
res <- pptomesh(X=spatstat.data::bei,
elev= spatstat.data::bei.extra$elev)
forest <- spforest(X = res)
log(forest+exp(-8))
}
