% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{merge.spforest}
\alias{merge.spforest}
\title{Merge two forests with same parameters}
\usage{
\method{merge}{spforest}(x, y, ...)
}
\arguments{
\item{...}{}
}
\description{
Merge two forests with same parameters
}
