% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{new_spforest}
\alias{new_spforest}
\title{Constructor for spforest}
\usage{
new_spforest(
  trees = list(),
  pt_intree = list(),
  X = ppp(),
  listcov = list(),
  p = double(),
  mtry = double()
)
}
\description{
Constructor for spforest
}
