% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OOBppmscr}
\alias{OOBppmscr}
\title{OOB ppm}
\usage{
OOBppmscr(forest, covariates, cores = 1)
}
\arguments{
\item{forest}{spforest}

\item{covariates}{list of covariates considered}
}
\value{
A number
}
\description{
OOB ppm
}
\examples{
forest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
}
