% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.spforest.R
\name{Ops.spforestmesh}
\alias{Ops.spforestmesh}
\title{Standard binary operations Ops}
\usage{
\method{Ops}{spforestmesh}(e1, e2 = NULL)
}
\arguments{
\item{e2}{}
}
\description{
Standard binary operations Ops
}
\examples{
unary operation
}
