% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{plot.sptree}
\alias{plot.sptree}
\title{Plot spatial intensity tree}
\usage{
\method{plot}{sptree}(x, ..., main)
}
\arguments{
\item{x}{A spatial intensity tree return by treerec function}

\item{...}{additional arguments}
}
\value{
A plot of the spatial intensity tree
and an \code{\link[spatstat]{im}} object.
}
\description{
Plot spatial intensity tree
}
\examples{
arbre <- treerec(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  minpts = 300,
  mtry = 1
)
plot(arbre)
}
