% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{predict.sptree_save}
\alias{predict.sptree_save}
\title{Tree prediction (Save when attempting another version)}
\usage{
\method{predict}{sptree_save}(object, newdata, ...)
}
\arguments{
\item{object}{A spatial intensity tree return by treerec function}

\item{newdata}{a xy vector or a ppp object}

\item{...}{Additional argument}
}
\value{
A number .....
}
\description{
Tree prediction (Save when attempting another version)
}
\examples{
arbre <- treerec(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  mtry = 1,
  minpts = 300
)
arbre$listcov <- list(
  grad = spatstat.data::bei.extra$grad,
  elev = spatstat.data::bei.extra$elev
)
predict(object = arbre, newdata = c(100, 100))
}
