% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predicttree}
\alias{predicttree}
\title{Tree prediction for importance}
\usage{
predicttree(object, newdata, ...)
}
\arguments{
\item{object}{A spatial intensity tree returned
by tesscovtree of tesstree functions}

\item{newdata}{a xy vector or a ppp object}

\item{...}{Additional argument}
}
\value{
A number .....
}
\description{
Tree prediction for importance
}
\details{
To compute the importance, we need to keep the structure
of a tree and not just the pixel image of the intensity that
it outputs.
}
\examples{
vecval0 <- lapply(beisoilres, FUN = function(i) {
  c(as.matrix.im(i))
})
arbre <- tesscovtree(
  X = spatstat.data::bei,
  vecval = vecval0,
  areapixel = beisoilres[[1]]$xstep * beisoilres[[1]]$ystep,
  dimcov = beisoilres[[1]]$dim,
  covrangex = beisoilres[[1]]$xrange,
  covrangey = beisoilres[[1]]$yrange,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500
)
predicttree(object = arbre, newdata = c(100, 100))
}
