% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{print.spforest}
\alias{print.spforest}
\title{Printing spatial intensity forest}
\usage{
\method{print}{spforest}(x, ...)
}
\arguments{
\item{x}{A spatial intensity forest return by RforestPP function}

\item{...}{Additional arguments}
}
\value{
A description of the random intensity forest with
the number of points, number of covariates, and number of trees used.
}
\description{
Printing spatial intensity forest
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
print(forest)
}
