% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{print.sptree}
\alias{print.sptree}
\title{Printing spatial intensity tree}
\usage{
\method{print}{sptree}(x, ...)
}
\arguments{
\item{x}{A spatial intensity tree return by treerec function}

\item{...}{Additional arguments}
}
\value{
A description of the random intensity tree with
the number of points and covariates used.
}
\description{
Printing spatial intensity tree
}
\examples{
arbre <- treerec(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  minpts = 300,
  mtry = 1
)
print(arbre)
}
