% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{rand_covar}
\alias{rand_covar}
\title{Randomise input for tree}
\usage{
rand_covar(listcovariates, mtry = 1)
}
\arguments{
\item{listcovariates}{list of covariatesd used}

\item{mtry}{selecting proba}
}
\value{
a vector of 0 and 1
}
\description{
Randomise input for tree
}
\examples{
rand_covar(
  listcovariates = list(1, 1, 1, 1, 1),
  mtry = 1 / 2
)
}
