% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newforest.R
\name{RforestPP2}
\alias{RforestPP2}
\title{Random Intensity Forest (improved)}
\usage{
RforestPP2(
  X,
  listcovariates = NULL,
  Ntree = 10,
  minpts = spatstat.geom::npoints(X)/10,
  mtry = 1/3,
  p = 0,
  cores_trees = 1,
  score = "lcv",
  threshold = smallest_pixelarea(listcovariates),
  tol = Inf,
  minsplitq = 0.5,
  maxsplitq = 0.5
)
}
\arguments{
\item{X}{A spatial point process as a
\code{\link[spatstat.geom]{ppp}} object from spatstat.}

\item{listcovariates}{A list of covariates as
\code{\link[spatstat.geom]{im}} objects from spatstat.}

\item{Ntree}{A positive integer.
The number of trees in the random intensity forest.}

\item{minpts}{A positive integer.
The minimum number of points in a region to allow a split.}

\item{mtry}{A number in \eqn{[0,1)}.
Probability that a covariate is used at each a split.}

\item{p}{A number in \eqn{[0,1)}.
Control the thinning process applied to the original point pattern \strong{X} before
fitting a tree intensity estimator.}

\item{cores_trees}{A positive integer.
The number of cores used to computes the intensity trees.}

\item{score}{String specifying the score used to choose among splits, see details.}

\item{threshold}{A positive number.
The minimum area of a region for which we allow at most one split.}

\item{tol}{unused.}

\item{minsplitq}{unused.}

\item{maxsplitq}{unused.}
}
\value{
An object of class \code{\link[Rsandbox]{spforest}}.
}
\description{
Compute a random intensity forest from an observed point pattern
and a list of covariates.
}
\details{
This function compute a random intensity forest using the covariates given
in \code{listcovariates}.
First the points of \code{X} are thinned with probability \code{p} if
\eqn{p>0} or bootstrapped if \eqn{p=0}.
Then the function call \code{\link{treerec}} to compute
an intensity tree estimate. This is repeated \code{Ntree} times.
The result returns in an \code{\link{spforest}} object.

When computing a random intensity, one need a stopping criterion after
which we do not try to split the domain of observation any more.
Two stopping criterion are implemented: based on the minimal area of the
window and on the minimal number of points.
When there is less than \code{minpts} in a cell,
we do not try to split the cell any more.
When a cell has an area below \code{threshold} we do not attempt to split.
In principle, this criterion can be used jointly but we have empirically
found that it was more convenient and similar in term of performances
to consider only \code{minpts} and by default, \code{threshold}
is the smallest area of a pixel in the covariates in \code{listcovariates}.
Note that it does not make sense
to have \code{threshold} smaller than the smallest pixel area in the covariates.

To increase the diversity between trees in random forest,
a common practice is to consider different covariates at each split.
This is controlled by \code{mtry}
which is the probability that a covariate, independently of the others,
is used at each split.

We let \eqn{n_1} and \eqn{n_2} be the number of points in the regions \eqn{W1},
\eqn{W2}, respectively. We let |W| be the area of a region W.
Currently, several scores implemented to compute the performance of a split:
\itemize{
\item "lcv", \eqn{\frac{n_1 \log{n_1-1}}{|W_1|}\mathbf{1}_{n_1>1} +
\frac{n_2\log{n_2-1}}{|W_2|} \mathbf{1}_{n_2>1}}
\item "lcv2", \eqn{\frac{n_1\log{n_1-1}}{|W_1|} +
\frac{n_2\log{n_2-1}}{|W_2|} + \infty \mathbf{1}_{n_1=1} + \infty \mathbf{1}_{n_2=1}}
\item "ent", \eqn{\frac{n_1\log{n_1}}{|W_1|}\mathbf{1}_{n_1>1} +
\frac{n_2\log{n_2}}{|W_2|} \mathbf{1}_{n_2>1}}
\item "star", \eqn{|\frac{n_1}{n_1+n_2} - \frac{|W_1|}{(|W_1| + |W_2|)}|}
\item "ise", \eqn{-\frac{n_1^2}{|W_1|} - \frac{n_2^2}{|W_2|}}
\item "isecv", \eqn{-\frac{n_1^2-2n_1}{|W_1|} - \frac{n_2^2-2n_2}{|W_2|}}
}
Empirically, there has been no score outputting better
integrated mean squares error than others.

The arguments \code{tol}, \code{minsplitq} and \code{maxsplitq}
are relics from the research development of the method and will most likely be dropped
in future versions.
}
\examples{
forest <- RforestPP2(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 100,
  mtry = 1 / 3,
  p = 0,
  cores_trees = 1
)
}
\references{
\url{Article arxiv version}
}
