% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_manifold.R
\name{sample_points_manifold}
\alias{sample_points_manifold}
\title{Title}
\usage{
sample_points_manifold(
  vertices,
  faces,
  triangle_centers,
  n,
  weights = FALSE,
  dimweight = 1
)
}
\arguments{
\item{dimweight}{}
}
\description{
Title
}
\examples{
Title

}
