% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_score.R
\name{score.pp}
\alias{score.pp}
\title{Score parent}
\usage{
score.pp(n0, W0area, score = "lcv")
}
\arguments{
\item{n0}{Number of points in point process}

\item{W0area}{Area of the ppp}

\item{score}{The score to use}
}
\value{
A number
}
\description{
Score parent
}
\examples{
score.split(5, 5)
}
