% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_score.R
\name{score.split}
\alias{score.split}
\title{Score of a split}
\usage{
score.split(n1, n2, W1area, W2area, score = "lcv")
}
\arguments{
\item{n1}{Number of points in left child}

\item{n2}{Number of points in right child}

\item{W1area}{Area of left child}

\item{W2area}{Area of right child}

\item{score}{The score to use}
}
\value{
A numeric number
}
\description{
It is used to compute the score.
}
\details{
Provide a lot of info blabla
}
\examples{
score.split(n1 = 5, n2 = 5, W1area = 5, W2area = 5)
}
