% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simppface}
\alias{simppface}
\title{Simulated point pattern on a 3D mesh}
\format{
A list of two elements: "mesh" and "pp".
}
\usage{
simppface
}
\description{
Contains a simulated inhomogeneous point pattern simulated on
the mesh humface from the R package Rvcg.
library(Rvcg)
library(rgl)
library(spforest)
data("humface")

features <- features_mesh(humface)
tricenter <- features$tricenter
nearest <- RANN::nn2(tricenter, rbind(humface$vb\link[=1:3]{, 7440}), k = 2000) # k = 1 => plus proche
meshspot <- humface
meshspot$it <- meshspot$it\link{, nearest$nn.idx}

set.seed(28)
X <- dummypponmesh(mesh = meshspot, n = 300, weights = FALSE)
Y <- dummypponmesh(mesh = humface, n = 850, weights = FALSE)
Z <- Y
Z$pp <- rbind(X$pp, Y$pp)

g <- function(pts) {
(0.2 * pts\link{1} - 0.9 * pts\link{2} + pts\link{3} + 70) < 0
}

idside <- sapply(1:dim(Z$pp)\link{1}, FUN = function(i) {
g(Z$pp\link{i, })
})

simppface <- list(mesh = Z$mesh, pp = Z$pp\link{!idside,})

save(simppface, file = "simppface.rda", compress = "xz")
}
\keyword{datasets}
