% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smallest_pixelarea}
\alias{smallest_pixelarea}
\title{Find the small pixel area in list of images}
\usage{
smallest_pixelarea(x)
}
\arguments{
\item{x}{an im object}
}
\value{
A number.
}
\description{
Find the small pixel area in list of images
}
\examples{
smallest_pixelarea(spatstat.data::bei.extra)
}
