% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{spforest.object}
\alias{spforest.object}
\title{spforest.object}
\description{
S3 class used to define random forest intensity
}
\details{
An object of this class represents a spatial intensity forest.
It contains
\itemize{
\item \code{imforest}, an pixel image as
\code{\link[spatstat.geom]{im.object}} object representing the value of
the estimated intensity on the window of \code{X}.
\item \code{trees}, a list of the spatial intensity trees as \code{\link{sptree.object}}.
\item \code{ntrees}, the number of trees in the forest.
\item \code{pt_intree}, a list where the \eqn{i}-th element
is a vector containing the index of the points of \code{X} used
to compute the \eqn{i}-th intensity tree.
\item \code{X}, the original point pattern as a \code{\link[spatstat.geom]{ppp}} object.
\item \code{listcov}, a list of the covariates used to compute the forest
as \code{\link[spatstat.geom]{im}} objects. If \code{NULL},
independently random tessellation have been used.
\item \code{p}, a number in \eqn{[0,1)}
controlling the thinning process applied to \code{X}
before computing a tree intensity estimator.
\item \code{mtry}, probability that a covariate is used at each a split.
}
}
