% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_cell.R
\name{splitcell}
\alias{splitcell}
\title{Splitting cell}
\usage{
splitcell(
  X,
  score = "lcv",
  listcovariates = NULL,
  usecovariates = rep(1, length(listcovariates)),
  thres.cell = 100,
  minpts = 0,
  tol = Inf,
  imp = NULL,
  minsplitq = 0.2,
  maxsplitq = 0.8
)
}
\arguments{
\item{X}{The observed data point pattern.}

\item{score}{A score to choose among "lcv", "lcv2", "ent", "star", "ise", "isecv".}

\item{listcovariates}{A list with all the covariates used for the tree. The
covariates must be given as im object from the package spatstat.}

\item{usecovariates}{Vector of 0,1 with covariate to choose}

\item{thres.cell}{Minimum size of a cell}

\item{minpts}{A positive integer. The minimum number of points allowed to try to split a cell.}

\item{tol}{A positive integer.
Look at the tol last gen to see if it was improved}

\item{imp}{A positive integer.
History of the improvement up to now. Used in random forest algorithm.}

\item{minsplitq}{A number in $(0,1)$. Determine the minimal allowed splitting value
of a covariable.}

\item{maxsplitq}{A number in $(0,1)$. Determine the maximal allowed splitting value
of a covariable.}
}
\value{
A list with PPleft = split.left[\link{id_best_scr}],
PPright = X[!leftlvl[\link{id_best_scr}]],
split_var = which(usecovariates == 1)\link{id_best_scr},
split_val = split_scrs\link{id_best_scr},
improvement = newimp,
scr_parent = scr_parent
}
\description{
Splitting cell
}
\examples{
splitcell(
  X = spatstat.data::bei,
  score = "lcv",
  listcovariates = list(
    grad = spatstat.data::bei.extra$grad,
    elev = spatstat.data::bei.extra$elev
  ),
  usecovariates = c(1, 1),
  thres.cell = 100,
  minpts = 10,
  tol = Inf,
  imp = NULL,
  minsplitq = 0.5,
  maxsplitq = 0.5
)
}
