% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{sptree.object}
\alias{sptree.object}
\title{sptree.object}
\description{
S3 class used to define random intensity trees.
}
\details{
An object of this class represents a spatial intensity tree.
It contains
\itemize{
\item \code{tree}, a list of nodes of the tree. The \eqn{i}-th
element contains a list with entries \code{nodeID, nX, left_daughter, 
right_daughter, split_var, split_val, whystop}.
\item \code{X}, the point pattern used to estimate the tree.
\item{namecov}, the name of the covariates used in
\code{listcovariates}.
\item{namelist}, TODO.
\item \code{listcov}, the list of covariates used.
\item \code{im}, a pixel image, as \code{\link[spatstat.geom]{im.object}},
representing the estimated intensity.
}
}
