% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{summary.sptree}
\alias{summary.sptree}
\title{Summary of a spatial intensity tree}
\usage{
\method{summary}{sptree}(object, fulltree = F, ...)
}
\arguments{
\item{object}{A spatial intensity tree return by tesstree function}

\item{fulltree}{Should we print all the column of the matrix?}

\item{...}{Additional arguments}
}
\value{
A dataframe representing each node and spit of the intensity tree.
}
\description{
Summary of a spatial intensity tree
}
\examples{
vecval0 <- lapply(beisoilres, FUN = function(i) {
  c(as.matrix.im(i))
})
arbre <- tesscovtree(
  X = spatstat.data::bei,
  vecval = vecval0,
  areapixel = beisoilres[[1]]$xstep * beisoilres[[1]]$ystep,
  dimcov = beisoilres[[1]]$dim,
  covrangex = beisoilres[[1]]$xrange,
  covrangey = beisoilres[[1]]$yrange,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500
)
summary(arbre)
}
