% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetess.R
\name{tessforest}
\alias{tessforest}
\title{Intensity tessellation forest}
\usage{
tessforest(
  X,
  lambda = NULL,
  Ntrees = 100,
  at = NULL,
  mc.cores = 1,
  test.connected = TRUE
)
}
\arguments{
\item{X}{ppp}

\item{lambda}{intensity}

\item{Ntrees}{number trees}

\item{at}{point at which we estimate}

\item{mc.cores}{how many cores}

\item{test.connected}{do we test if the voronoi cells are connected}
}
\value{
The estimated intensity as an \code{\link[spatstat.geom]{im}}.
}
\description{
Intensity tessellation forest
}
\examples{
tessforest(X = bei, lambda = 100, Ntree = 5, test.connected = FALSE)
}
