% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtreetess.R
\name{tessforest2}
\alias{tessforest2}
\title{Intensity tessellation forest}
\usage{
tessforest2(
  X,
  Ntree = 1,
  lambda = 100,
  dimyx = c(50, 50),
  test.connected = FALSE,
  cores = 1
)
}
\arguments{
\item{X}{The observed data point pattern,
as a \code{\link[spatstat.geom]{ppp}} object .}

\item{Ntree}{A positive integer.
The number of trees in the random intensity forest.}

\item{lambda}{An integer. The number of points used for random tessellation.}

\item{dimyx}{A vector of two integers. The dimensions of the output image passed
to \code{\link[spatstat.geom]{as.im}}.}

\item{test.connected}{Logical. If \code{TRUE},
\code{\link[spatstat.geom]{connected}} is applied to the tessellation to split tiles
in different connected components. It is only useful if the windows of
observation od \code{X} is not convex.}

\item{cores}{A positive integer.
The number of cores used to computes the intensity trees.}
}
\value{
A pixel image, object of class \code{\link[spatstat.geom]{im.object}}.
}
\description{
Intensity tessellation forest
}
\examples{
Z <- tessforest2(
  X = bei,
  lambda = 100,
  dimyx = c(101, 201),
  test.connected = FALSE,
  Ntree = 5,
  cores = 1
)
plot(Z)
}
