% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetess.R
\name{tesstree}
\alias{tesstree}
\title{Intensity tesselation tree}
\usage{
tesstree(X, lambda, target.points, test.connected = TRUE)
}
\arguments{
\item{test.connected}{}
}
\description{
Intensity tesselation tree
}
