% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetess.R
\name{tesstree_precision}
\alias{tesstree_precision}
\title{Intensity tessellation tree high definition}
\usage{
tesstree_precision(X, lambda, target.points, test.connected = TRUE)
}
\arguments{
\item{X}{a ppp}

\item{lambda}{intensity}

\item{target.points}{points at which we estimated the intensity}

\item{test.connected}{do we test if the voronoi cells are connected}
}
\value{
image of the intensity ? To check
}
\description{
Intensity tessellation tree high definition
}
\examples{
tesstree_precision(rpoispp(10), lambda = 50, target.points = rpoispp(10))
}
