% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{treerec}
\alias{treerec}
\title{Spatial Intensity tree function}
\usage{
treerec(
  X,
  score = "lcv",
  threshold = spatstat.geom::area(X)/10000,
  listcovariates = NULL,
  mtry = 1,
  tol = Inf,
  minpts = spatstat.geom::npoints(X)/10,
  minsplitq = 0.5,
  maxsplitq = 0.5,
  inforest = F
)
}
\arguments{
\item{X}{The observed data point pattern.}

\item{score}{A score to choose among "lcv", "lcv2", "ent", "star", "ise", "isecv".}

\item{threshold}{minimum threshold to allow to split cell}

\item{listcovariates}{A list with all the covariates used for the tree. The
covariates must be given as im object from the package spatstat.}

\item{mtry}{Probability of choosing a covariable}

\item{tol}{A positive integer.
Look at the tol last gen to see if it was improved}

\item{minpts}{A positive integer. The minimum number of points allowed to try to split a cell.}

\item{minsplitq}{A number in $(0,1)$. Determine the minimal allowed splitting value
of a covariable.}

\item{maxsplitq}{A number in $(0,1)$. Determine the maximal allowed splitting value
of a covariable.}
}
\description{
Spatial Intensity tree function
}
\examples{
arbre <- treerec(
  X = spatstat.data::bei,
  threshold = 1000,
  score = "lcv",
  listcovariates = list(
    grad = spatstat.data::bei.extra$grad,
    elev = spatstat.data::bei.extra$elev
  ),
  mtry = 1,
  tol = Inf,
  minpts = 50,
  minsplitq = 0.5,
  maxsplitq = 0.5
)
}
