% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{vipplot}
\alias{vipplot}
\title{Vip barplot of forest}
\usage{
vipplot(x, sorted = F, cores = 1, ...)
}
\arguments{
\item{x}{An spforest object}

\item{cores}{To compute faster}

\item{...}{why not}
}
\description{
Vip barplot of forest
}
