% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_manifold.R
\name{add_colorbar3d}
\alias{add_colorbar3d}
\title{Title}
\usage{
add_colorbar3d(
  center,
  length = 1,
  width = 0.05,
  theta = 0,
  phi = 0,
  zlim = c(0, 1),
  ncolors = 100,
  colfun = colorRampPalette(c("blue", "red")),
  nticks,
  lasttick = TRUE,
  title = NULL,
  cex = 1
)
}
\arguments{
\item{cex}{}
}
\description{
Title
}
