% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{boxplot.spforest}
\alias{boxplot.spforest}
\title{Boxplot forest}
\usage{
\method{boxplot}{spforest}(x, cores = 1, ...)
}
\arguments{
\item{x}{the forest}

\item{cores}{To compute faster}

\item{...}{ignored.}
}
\value{
Boxplot of the variable importances.
}
\description{
Boxplot forest
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
boxplot(forest)
}
