% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{boxplot.spforest}
\alias{boxplot.spforest}
\title{Boxplot forest}
\usage{
boxplot(x = forest)
}
\arguments{
\item{x}{A \code{\link{spforest.object}} with \code{listcov} not \code{NULL}.}

\item{viptype}{An integer in \{1,2,3,4\} passed to \code{\link[spforest]{importance}}.}

\item{...}{Arguments passed to \code{\link[graphics]{boxplot}}.}
}
\value{
A list returned by the function \code{\link[graphics]{boxplot}}.
}
\description{
Plot the boxplot of the importance of each covariate
in each tree of the random forest intensity
}
\details{
First the function \code{\link[spforest]{importance}} is called to
compute the importance of each covariate in each tree \code{x}.
Then, a boxplot is drawn to visualize the distribution of the
importance of each covariate.
}
\examples{
forest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
boxplot(forest)
}
