% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{extractforest}
\alias{extractforest}
\title{Extract a smaller forest}
\usage{
extractforest(forest, whichtrees)
}
\arguments{
\item{forest}{A \code{\link{spforest}[spforest.object]} .}

\item{whichtrees}{A vector indicating which trees to use in the
new forest.}
}
\value{
A \code{\link{spforest}[spforest.object]} .
}
\description{
Given a random forest intensity estimate, return a random forest intensity estimated
with a only subset of all the trees.
}
\details{
The function extracts the trees indicated in \code{whichtrees}
from \code{forest} and returns a new \code{\link{spforest}[spforest.object]}.
The intensity estimate of the new forest is obtained by averaging the intensity
estimates of the selected trees. All the other arguments are updated accordingly.
}
\examples{
bigforest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
smallforest <- extractforest(
  forest = bigforest,
  whichtrees = c(2, 5)
)
}
