% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{importance}
\alias{importance}
\title{Importance of one covariable}
\usage{
importance(forest, id_cov, cores = 1)
}
\arguments{
\item{forest}{A spforest object}

\item{id_cov}{The id in forest$listcov of the covariable looked at.}

\item{cores}{how many cores to use to speed up computation}
}
\description{
Importance of one covariable
}
