% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{imspforest}
\alias{imspforest}
\title{Compute image of a spatial intensity forest}
\usage{
imspforest(x, ...)
}
\arguments{
\item{x}{A spatial intensity tree return by RforestPP function}

\item{...}{additional arguments}
}
\description{
Compute image of a spatial intensity forest
}
