% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtreerec.R
\name{intensitytree}
\alias{intensitytree}
\title{Spatial Intensity tree function (improved)}
\usage{
intensitytree(
  X,
  listcovariates,
  minpts = 500,
  mtry = 1,
  score = "lcv",
  threshold = spatstat.geom::area(X)/10000,
  inforest = F
)
}
\arguments{
\item{X}{The observed data point pattern.}

\item{listcovariates}{A list with all the covariates used for the tree. The
covariates must be given as im object from the package spatstat.}

\item{minpts}{A positive integer.
The minimum number of points allowed to try to split a cell.}

\item{mtry}{Probability of choosing a covariable.}

\item{score}{A score to choose among "lcv", "lcv2", "ent", "star", "ise", "isecv".}

\item{threshold}{Minimum threshold to allow to split cell.}

\item{inforest}{Logical. Indicates if the function is run in a forest or not.}
}
\value{
An object of class \code{sptree}.
}
\description{
Spatial Intensity tree function (improved)
}
\examples{
mytree <- intensitytree(
  X = spatstat.data::bei,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500
)
plot(mytree)
}
