% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.spforest.R
\name{Math.spforestmesh}
\alias{Math.spforestmesh}
\title{Dispatch standard math functions}
\usage{
\method{Math}{spforestmesh}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Dispatch standard math functions
}
