% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{merge.spforest}
\alias{merge.spforest}
\title{Merge two forests with same parameters}
\usage{
\method{merge}{spforest}(x, y, ...)
}
\arguments{
\item{x, y}{Two \code{\link{spforest}[spforest.object]} with identical entries
\code{p}, \code{mtry}, \code{listcov} and \code{X}.}

\item{...}{Ignored}
}
\value{
An \code{\link[spforest]{spforest.object}}
combining the information of \code{x} and \code{y}.
}
\description{
Given two random forest intensity estimates,
merge them into one \code{\link{spforest}[spforest.object]}.
}
\details{
The function merges two \code{\link{spforest}[spforest.object]} by concatenating
their entires \code{trees}, \code{pt_intree} and
averaging their intensity estimates given by their entries \code{imforest}.
If any of the entries \code{p}, \code{mtry}, \code{listcov} or \code{X} is different,
the function returns an error.
}
\examples{
forest1 <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
forest2 <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 5,
  minpts = 300,
  mtry = 1
)
merge(forest1, forest2)
}
