% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{merge.spforest}
\alias{merge.spforest}
\title{Merge two forests with same parameters}
\usage{
\method{merge}{spforest}(x, y, ...)
}
\arguments{
\item{x}{First forest}

\item{y}{Second forest}

\item{...}{ignored}
}
\value{
An \code{\link{spforest}} object
the intensity trees from both forest.
}
\description{
Merge two forests with same parameters
}
\examples{
forest1 <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
forest2 <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
merge(forest1, forest2)
}
