% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OOBoptim}
\alias{OOBoptim}
\title{Find best OOB parameter combination}
\usage{
OOBoptim(X, listcovariates, params, ...)
}
\arguments{
\item{X}{The point pattern as a \code{\link[spatstat.geom]{ppp}} .}

\item{listcovariates}{list of covariates}

\item{params}{A list containing entries mtry, minpts and, optionally, Ntree.}

\item{...}{Other arguments passed to spforest.}
}
\value{
A dataframe with 3 columns: mtry, minpts and OOB
}
\description{
Find best OOB parameter combination
}
\details{
A spatial random forest is generated from the combinations of all the
arguments given in params. If Ntree is not given, it is set to 50 by defaults.
The arguments ... are passed to the function \code{\link{spforest}}.
}
\examples{
X = spatstat.data::bei
listcovariates = spatstat.data::bei.extra
params = list(mtry = c(1,2), minpts = c(50,100,200))
OOBoptim(X = X, listcovariates = listcovariates, Ntree = 50, params = params)
}
