% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{OOBppmscr}
\alias{OOBppmscr}
\title{OOB ppm}
\usage{
OOBppmscr(forest, covariates, cores = 1)
}
\arguments{
\item{forest}{spforest}

\item{covariates}{list of covariates considered}

\item{cores}{to spped up}
}
\description{
OOB ppm
}
