% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OOBscr}
\alias{OOBscr}
\title{OOB forest}
\usage{
OOBscr(forest, cores = 1)
}
\arguments{
\item{forest}{spforest}

\item{cores}{to speed up}
}
\value{
A number.
}
\description{
OOB forest
}
\examples{
forest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
OOBscr(forest, cores = 1)
}
