% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.spforest.R
\name{Ops.spforestmesh}
\alias{Ops.spforestmesh}
\title{Standard binary operations Ops}
\usage{
\method{Ops}{spforestmesh}(e1, e2 = NULL)
}
\arguments{
\item{e1}{An object of class \code{spforestmesh}}

\item{e2}{Either a numeric or an object of class \code{spforestmesh}}
}
\value{
An object of class \code{spforestmesh}
}
\description{
Standard binary operations Ops
}
\examples{
res <- pptomesh(X=spatstat.data::bei,
elev= spatstat.data::bei.extra$elev)
forest <- spforest(X = res)
forest + forest
unary operation
}
