% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{plot.spforest}
\alias{plot.spforest}
\title{Plot spatial intensity forest}
\usage{
\method{plot}{spforest}(x, ..., main)
}
\arguments{
\item{x}{A spatial intensity tree return by RforestPP function}

\item{...}{additional arguments}
}
\value{
Same as \code{\link[spatstat.geom]{plot.im}}.
}
\description{
Plot spatial intensity forest
}
\details{
This function first convert an \code{\link{spforest}}
as an \code{\link[spatstat.geom]{im}} object and then plot it with
\code{\link[spatstat.geom]{plot.im}}. All arguments in \code{...} are passed to
\code{\link[spatstat.geom]{plot.im}}.
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
plot.spforest(forest)
}
