% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{plot.spforest}
\alias{plot.spforest}
\title{Plot spatial intensity forest}
\usage{
\method{plot}{spforest}(x, ..., main)
}
\arguments{
\item{x}{A spatial intensity tree return by RforestPP function}

\item{...}{additional arguments}
}
\description{
Plot spatial intensity forest
}
