% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{plot.spforestmesh}
\alias{plot.spforestmesh}
\title{Plot spatial intensity forest on a manifold}
\usage{
\method{plot}{spforestmesh}(x, points = FALSE, colorbar = FALSE, log = F, offset = exp(-8), ...)
}
\arguments{
\item{x}{A spforestmesh object}

\item{...}{additional arguments}

\item{main}{A title for the plot.}
}
\description{
Plot a \code{\link{spforest.object}}.
}
\details{
details
}
