% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{plot.sptree}
\alias{plot.sptree}
\title{Plot spatial intensity tree}
\usage{
\method{plot}{sptree}(x, ..., main = "Spatial Intensity Tree")
}
\arguments{
\item{x}{A spatial intensity tree return by tesstree function}

\item{...}{additional arguments}

\item{main}{Title of the plot.}
}
\value{
A plot of the spatial intensity tree
and an \code{\link[spatstat.geom]{im}} object.
}
\description{
Plot spatial intensity tree
}
\examples{
vecval0 <- lapply(beisoilres, FUN = function(i) {
  c(as.matrix.im(i))
})
arbre <- tesscovtree(
  X = spatstat.data::bei,
  vecval = vecval0,
  areapixel = beisoilres[[1]]$xstep * beisoilres[[1]]$ystep,
  dimcov = beisoilres[[1]]$dim,
  covrangex = beisoilres[[1]]$xrange,
  covrangey = beisoilres[[1]]$yrange,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500
)
plot(arbre)
}
