% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{plot.sptree}
\alias{plot.sptree}
\title{Plot spatial intensity tree}
\usage{
\method{plot}{sptree}(x, ...)
}
\arguments{
\item{x}{A spatial intensity tree return by treerec function}

\item{...}{additional arguments}
}
\description{
Plot spatial intensity tree
}
