% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_manifold.R
\name{plot_manifold_intensity}
\alias{plot_manifold_intensity}
\title{Title}
\usage{
plot_manifold_intensity(
  forestmesh,
  points = FALSE,
  size = 2,
  colorbar = FALSE,
  theta = 10,
  phi = -50,
  zoom = 0.8,
  pos = NULL,
  nticks = 5,
  lasttick = TRUE
)
}
\arguments{
\item{size}{description}

\item{lasttick}{}
}
\description{
Title
}
