% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{predict.spforest}
\alias{predict.spforest}
\title{Forest prediction}
\usage{
\method{predict}{spforest}(object, newdata, ...)
}
\arguments{
\item{object}{A spatial intensity forest return by RforestPP function}

\item{newdata}{a xy vector or a ppp object}

\item{...}{Additional argument}
}
\value{
A number .....
}
\description{
Forest prediction
}
