% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{predict.spforest}
\alias{predict.spforest}
\title{Forest prediction}
\usage{
\method{predict}{spforest}(object, newdata, ...)
}
\arguments{
\item{object}{A spatial intensity forest return by RforestPP function}

\item{newdata}{a xy vector or a ppp object}

\item{...}{Additional argument}
}
\value{
A vector of predicted intensity
}
\description{
Forest prediction
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
predict(forest, c(100, 100))
}
