% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{predict.sptree}
\alias{predict.sptree}
\title{Tree prediction}
\usage{
\method{predict}{sptree}(object, newdata, ...)
}
\arguments{
\item{object}{A spatial intensity tree returned
by tesscovtree of tesstree functions}

\item{newdata}{a xy vector or a ppp object}

\item{...}{Additional argument}
}
\value{
A number .....
}
\description{
Tree prediction
}
\examples{
vecval0 <- lapply(beisoilres, FUN = function(i) {
  c(as.matrix.im(i))
})
arbre <- tesscovtree(
  X = spatstat.data::bei,
  vecval = vecval0,
  areapixel = beisoilres[[1]]$xstep * beisoilres[[1]]$ystep,
  dimcov = beisoilres[[1]]$dim,
  covrangex = beisoilres[[1]]$xrange,
  covrangey = beisoilres[[1]]$yrange,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500
)
predict(object = arbre, newdata = c(100, 100))
}
