% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{print.spforest}
\alias{print.spforest}
\title{Printing spatial intensity forest}
\usage{
\method{print}{spforest}(x, ...)
}
\arguments{
\item{x}{A spatial intensity forst return by RforestPP function}

\item{...}{Additional arguments}
}
\description{
Printing spatial intensity forest
}
