% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{print.sptree}
\alias{print.sptree}
\title{Printing spatial intensity tree}
\usage{
\method{print}{sptree}(x, ...)
}
\arguments{
\item{x}{A spatial intensity tree return by treerec function}

\item{...}{Additional arguments}
}
\description{
Printing spatial intensity tree
}
