% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{print.sptree}
\alias{print.sptree}
\title{Printing spatial intensity tree}
\usage{
\method{print}{sptree}(x, ...)
}
\arguments{
\item{x}{A spatial intensity tree return by tesstree function}

\item{...}{Additional arguments}
}
\value{
A description of the random intensity tree with
the number of points and covariates used.
}
\description{
Printing spatial intensity tree
}
\examples{
vecval0 <- lapply(beisoilres, FUN = function(i) {
  c(as.matrix.im(i))
})
arbre <- tesscovtree(
  X = spatstat.data::bei,
  vecval = vecval0,
  areapixel = beisoilres[[1]]$xstep * beisoilres[[1]]$ystep,
  dimcov = beisoilres[[1]]$dim,
  covrangex = beisoilres[[1]]$xrange,
  covrangey = beisoilres[[1]]$yrange,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500
)
print(arbre)
}
