% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{RforestPP}
\alias{RforestPP}
\title{Title}
\usage{
RforestPP(
  X,
  listcovariates = NULL,
  score = "lcv",
  p = 0.8,
  Ntree = 10,
  threshold = spatstat.geom::area(X)/2^4,
  cores_trees = 1,
  mtry = 1/3,
  tol = Inf,
  minpts = 0,
  minsplitq = 0.5,
  maxsplitq = 0.5
)
}
\arguments{
\item{X}{XXX}

\item{listcovariates}{XXX}

\item{score}{XXX}

\item{p}{XXX}

\item{Ntree}{XXX}

\item{threshold}{XXX}

\item{cores_trees}{XXX}

\item{mtry}{XXX}

\item{tol}{XXX}

\item{minpts}{XXX}

\item{minsplitq}{XXX}

\item{maxsplitq}{XXX}
}
\description{
Title
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = list(
    grad = spatstat.data::bei.extra$grad,
    elev = spatstat.data::bei.extra$elev
  ),
  score = "lcv2",
  p = 1,
  Ntree = 3,
  threshold = spatstat.geom::area(spatstat.data::bei) / 2^4,
  cores_trees = 1,
  mtry = 1 / 3,
  tol = Inf,
  minpts = 50,
  minsplitq = 0.5,
  maxsplitq = 0.5
)
}
