% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{RforestPP}
\alias{RforestPP}
\title{Title}
\usage{
RforestPP(
  X,
  listcovariates = NULL,
  score = "lcv",
  p = 0,
  Ntree = 10,
  threshold = smallest_pixelarea(listcovariates),
  cores_trees = 1,
  mtry = 1/3,
  tol = Inf,
  minpts = spatstat.geom::npoints(X)/10,
  minsplitq = 0.5,
  maxsplitq = 0.5
)
}
\arguments{
\item{X}{A spatial point process \code{\link[spatstat.geom]{compatible.im}} \link{ppp} as  "ppp" object from spatstat.}

\item{listcovariates}{A list of covariates as "im" objects from spatstat.}

\item{score}{String specifying the score used to choose among splits, see details.}

\item{p}{Numeric. Control the thinning of the data applied to fit each tree.}

\item{Ntree}{Numeric. The number of trees in the random intensity forest.}

\item{threshold}{Numeric. The minimum area of a region for which we allow at most one split.}

\item{cores_trees}{Numeric. The number of cores used to computes the intensity trees.}

\item{mtry}{Numeric. Probability that a covariate is used at each a split.}

\item{tol}{unused}

\item{minpts}{Numeric. The minimum number of points in a region to allow a split.}

\item{minsplitq}{unused}

\item{maxsplitq}{unused \link{compatible.im}}
}
\description{
Title
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = list(
    grad = spatstat.data::bei.extra$grad,
    elev = spatstat.data::bei.extra$elev
  ),
  score = "lcv",
  p = 1,
  Ntree = 3,
  threshold = spatstat.geom::area(spatstat.data::bei) / 2^4,
  cores_trees = 1,
  mtry = 1 / 3,
  tol = Inf,
  minpts = 50,
  minsplitq = 0.5,
  maxsplitq = 0.5
)
}
