% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_manifold.R
\name{sample_points_manifold}
\alias{sample_points_manifold}
\title{Title}
\usage{
sample_points_manifold(
  tricenter,
  vertices,
  faces,
  n,
  weights = TRUE,
  dimweight = 3
)
}
\arguments{
\item{faces}{description}

\item{dimweight}{}
}
\description{
Title
}
