% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_score.R
\name{score.pp}
\alias{score.pp}
\title{Score parent}
\usage{
score.pp(X, score = "lcv")
}
\arguments{
\item{X}{A ppp object for the parent}

\item{score}{The score used}
}
\value{
A number
}
\description{
Score parent
}
\examples{
score.pp(X = spatstat.random::rpoispp(10))
}
