% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{spforest}
\alias{spforest}
\title{Random intensity forest}
\usage{
spforest(
  X,
  listcovariates = NULL,
  Ntree = 10,
  minpts = spatstat.geom::npoints(X)/10,
  mtry = 1/3,
  randmtry = FALSE,
  p = 0,
  score = "lcv",
  threshold = 2 * smallest_pixelarea(listcovariates),
  gamma = NULL,
  dimyx = c(50, 50),
  test.connected = FALSE,
  cores = 1
)
}
\arguments{
\item{X}{A spatial point process as a \code{\link[spatstat.geom]{ppp}} object.}

\item{listcovariates}{A list of covariates as \code{\link[spatstat.geom]{im}} objects.}

\item{Ntree}{Number of trees in the forest.}

\item{minpts}{A positive integer.
The minimum number of points after which we try to split a cell one last time.}

\item{mtry}{Probability of choosing a covariate.}

\item{randmtry}{Logical. If \code{TRUE}, \code{mtry} must be between 0 and 1 and
represents the probability to use each covariate at each split. If \code{FALSE}, \code{mtry}
covariates are randomly chosen at each split.}

\item{p}{A number in \eqn{[0,1)}
controlling the thinning process applied to \code{X}.
See \code{\link[spforest]{tesscovforest}} for details.}

\item{score}{A score to choose among "lcv", "lcv2", "ent", "star", "ise", "isecv".}

\item{threshold}{Minimum area to allow to split cell.}

\item{gamma}{An integer. The number of points used for random tessellation.}

\item{dimyx}{A vector of two integers. The dimensions of the output image passed
to \code{\link[spatstat.geom]{as.im}}.}

\item{test.connected}{Logical. If \code{TRUE},
\code{\link[spatstat.geom]{connected}} is applied to the tessellation to split tiles
in different connected components. It is only useful if the windows of
observation of \code{X} is not convex.}

\item{cores}{A positive integer. The number of cores to use.
If strictly larger than 1, parallel computing
is used to dispatch each intensity tree computation on different cores.}
}
\value{
An object of class \code{\link{spforest}}.
}
\description{
Estimate the intensity of a spatial point process by random intensity forest.
}
\details{
If \code{listcovariates} is not \code{NULL},
the function computes a random intensity forest using
\code{\link{tesscovforest}}. Otherwise, it computes a random intensity forest
using \code{\link{tessforest}}.
All arguments are passed to the corresponding function.
}
\examples{
forestwithcov <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 5,
  minpts = 200,
  mtry = 2
)
plot(forestwithcov)
forestwithtoutcov <- tessforest(
  X = bei,
  gamma = 50,
  dimyx = c(101, 201),
  Ntree = 100
)
plot(forestwithtoutcov)
}
