% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_cell.R
\name{splitcell}
\alias{splitcell}
\title{Splitting cell}
\usage{
splitcell(
  X,
  valpts,
  vecval,
  usecovariates,
  areapixel,
  score = "lcv",
  threshold = spatstat.geom::area(X)/10000
)
}
\arguments{
\item{X}{The observed data point pattern.}

\item{valpts}{A list. Values of each covariates at the points of \code{X}.}

\item{vecval}{A list. Values of the covariates at each pixel.}

\item{usecovariates}{Vector of 0 and 1 values indicating which covariates
to choose}

\item{areapixel}{The pixel area used in each covariates.}

\item{score}{A score to choose among "lcv", "lcv2", "ent", "star", "ise", "isecv".}

\item{threshold}{Minimum threshold to allow to split cell.
\code{\link[spatstat.geom]{im}}.}
}
\value{
A list with  split_var = split_var,
split_val = split_val,
sublevels = sublevels,
nsub = nsub,
suplevels = suplevels,
nsup = nsup,
whystop = NULL
}
\description{
Splitting cell
}
\examples{
listcovariates <- Rsandbox::beisoilres
X <- bei
valpts <- lapply(listcovariates,
  FUN = function(i) {
    i[X]
  }
)
vecval <- lapply(Rsandbox::beisoilres, FUN = function(i) {
  if (!spatstat.geom::is.im(i)) {
    stop("Elements of listcovar must be an im object")
  }
  c(as.matrix.im(i))
})
usecovariates <- rep(1, 15)
areapixel <- listcovariates[[1]]$xstep * listcovariates[[1]]$ystep
score <- "lcv"
A <- splitcell(
  X = X,
  valpts = valpts,
  vecval = vecval,
  usecovariates = usecovariates,
  areapixel = areapixel,
  threshold = 100
)
}
