% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{summary.sptree}
\alias{summary.sptree}
\title{Summary of a spatial intensity tree}
\usage{
\method{summary}{sptree}(object, fulltree = F, ...)
}
\arguments{
\item{object}{A spatial intensity tree return by treerec function}

\item{fulltree}{Should we print all the column of the matrix?}

\item{...}{Additional arguments}
}
\value{
A dataframe representing each node and spit of the intensity tree.
}
\description{
Summary of a spatial intensity tree
}
\examples{
arbre <- treerec(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  minpts = 300,
  mtry = 1
)
summary(arbre)
}
