% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{summary.sptree}
\alias{summary.sptree}
\title{Summary of a spatial intensity tree}
\usage{
\method{summary}{sptree}(object, fulltree = F, ...)
}
\arguments{
\item{object}{A spatial intensity tree return by treerec function}

\item{fulltree}{Should we print all the column of the matrix?}

\item{...}{Additional arguments}
}
\description{
Summary of a spatial intensity tree
}
