% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesstree.R
\name{tesscovtree}
\alias{tesscovtree}
\title{Spatial Intensity tree function with covariates}
\usage{
tesscovtree(
  X,
  vecval,
  areapixel,
  dimcov,
  covrangex,
  covrangey,
  listcovariates,
  minpts = 500,
  minarea = 100,
  mtry = 1,
  randmtry = FALSE,
  score = "lcv",
  threshold = spatstat.geom::area(X)/10000,
  inforest = F
)
}
\arguments{
\item{X}{The observed data point pattern.}

\item{vecval}{A list. The i-th element is the values of
a i-th covariate in \code{listcovariates} at the points of \code{X}.}

\item{areapixel}{The pixel area used for all covariate.}

\item{dimcov}{The element \code{dim} of the covariates which are
\code{\link[spatstat.geom]{im}}.}

\item{covrangex}{The element \code{xrange} of the covariates which are
\code{\link[spatstat.geom]{im}}.}

\item{covrangey}{The element \code{yrange} of the covariates which are
\code{\link[spatstat.geom]{im}}.}

\item{listcovariates}{A list with all the covariates used for the tree. The
covariates must be given as im object from the package spatstat.}

\item{minpts}{A positive integer.
The minimum number of points allowed to try to split a cell.}

\item{minarea}{A positive integer.}

\item{mtry}{Probability of choosing a covariate.}

\item{randmtry}{Logical. If \code{TRUE}, \code{mtry} must be between 0 and 1 and
represents the probability to use each covariate at each split. If \code{FALSE}, \code{mtry}
covariates are randomly chosen at each split.}

\item{score}{A score to choose among "lcv", "lcv2", "ent", "star", "ise", "isecv".}

\item{threshold}{Minimum threshold to allow to split cell.}

\item{inforest}{Logical. Indicates if the function is run in a forest or not.}
}
\value{
A pixel image as
\code{\link[spatstat.geom]{im.object}}.
}
\description{
Spatial Intensity tree function with covariates
}
\details{
The function generates a spatial intensity
tree from a point pattern \code{X} using a tessellation
build from covariates. The covariates needs to be
input in a vectorised form in \code{vecval}.

The tessellation is build iteratively, starting from
full window of \code{X} as the first cell, until there is less
than \code{minpts} points in a cell or
the area of the cell is less that \code{threshold}.

At each step, we look to split the cell
according to the sub and sup level sets of
the vectorised covariate in \code{vecval}.
We choose among them the split that
gives the maximal score, given by \code{score}.
Moreover, at each step, only a randomly subset of
to covariates, each selected with probability \code{mtry}
are used.

The arguments \code{dimcov}, \code{covrangex} and \code{covrangey},
are passed to \code{\link[spatstat.geom]{im}} to return
a pixel image of the spatial intensity tree.

The function output an object of class  \code{sptree} which
contains the point pattern \code{X}.
This is useful for information purposes.
However, if the function is run
as part of an intensity tessellation forest,
it uses a lot of memory to create many copies of \code{X}.
To avoid this, set \code{inforest=T}.
}
\examples{
vecval0 <- lapply(beisoilres, FUN = function(i) {
  c(as.matrix.im(i))
})
mytree <- tesscovtree(
  X = spatstat.data::bei,
  vecval = vecval0,
  areapixel = beisoilres[[1]]$xstep * beisoilres[[1]]$ystep,
  dimcov = beisoilres[[1]]$dim,
  covrangex = beisoilres[[1]]$xrange,
  covrangey = beisoilres[[1]]$yrange,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500,
  minarea = 20
)
plot(mytree)
}
