% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesstree.R
\name{tesscovtree}
\alias{tesscovtree}
\title{Spatial Intensity tree function with covariates}
\usage{
tesscovtree(
  X,
  vecval,
  areapixel,
  dimcov,
  covrangex,
  covrangey,
  listcovariates,
  minpts = 500,
  mtry = 1,
  score = "lcv",
  threshold = spatstat.geom::area(X)/10000,
  inforest = F
)
}
\arguments{
\item{X}{The observed data point pattern.}

\item{vecval}{A list. The i-th element is the values of a i-th covariate in
\code{listcovariates} at the points of \code{X}.}

\item{areapixel}{The pixel area used for all covariate.}

\item{dimcov}{The element \code{dim} of the covariates which are
\code{\link[spatstat.geom]{im}}.}

\item{covrangex}{The element \code{xrange} of the covariates which are
\code{\link[spatstat.geom]{im}}.}

\item{covrangey}{The element \code{yrange} of the covariates which are
\code{\link[spatstat.geom]{im}}.}

\item{listcovariates}{A list with all the covariates used for the tree. The
covariates must be given as im object from the package spatstat.}

\item{minpts}{A positive integer.
The minimum number of points allowed to try to split a cell.}

\item{mtry}{Probability of choosing a covariable.}

\item{score}{A score to choose among "lcv", "lcv2", "ent", "star", "ise", "isecv".}

\item{threshold}{Minimum threshold to allow to split cell.}

\item{inforest}{Logical. Indicates if the function is run in a forest or not.}
}
\value{
An object of class \code{sptree}.
}
\description{
Spatial Intensity tree function with covariates
}
\examples{
vecval0 <- lapply(beisoilres, FUN = function(i) {
  c(as.matrix.im(i))
})
mytree <- tesscovtree(
  X = spatstat.data::bei,
  vecval = vecval0,
  areapixel = beisoilres[[1]]$xstep * beisoilres[[1]]$ystep,
  dimcov = beisoilres[[1]]$dim,
  covrangex = beisoilres[[1]]$xrange,
  covrangey = beisoilres[[1]]$yrange,
  listcovariates = beisoilres,
  mtry = 1,
  minpts = 500
)
plot(mytree)
}
