% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetess.R
\name{tessforest}
\alias{tessforest}
\title{Intensity tesselation forest}
\usage{
tessforest(
  X,
  lambda = NULL,
  Ntrees = 100,
  at = NULL,
  mc.cores = 6,
  test.connected = TRUE
)
}
\arguments{
\item{test.connected}{}
}
\description{
Intensity tesselation forest
}
