% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessforest.R
\name{tessforest}
\alias{tessforest}
\title{Intensity tessellation forest without covariate}
\usage{
tessforest(
  X,
  Ntree = 1,
  gamma = 100,
  dimyx = c(50, 50),
  test.connected = FALSE,
  verbose = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{X}{The observed data point pattern,
as a \code{\link[spatstat.geom]{ppp.object}}.}

\item{Ntree}{A positive integer.
The number of trees in the random intensity forest.}

\item{gamma}{An integer. The number of points used to generate random tessellation.}

\item{dimyx}{A vector of two integers representing the dimensions of the output image passed
to \code{\link[spatstat.geom]{as.im}}.}

\item{test.connected}{Logical. If \code{TRUE},
\code{\link[spatstat.geom]{connected}} is applied to the tessellation to split tiles
in different connected components. It is only useful if the windows of
observation of \code{X} is not convex.}

\item{verbose}{If TRUE, display progress bar.}

\item{parallel}{Check if future::plan has been called.}
}
\value{
A pixel image, object of class \code{\link[spatstat.geom]{im.object}}.
}
\description{
Intensity tessellation forest without covariate
}
\details{
This function compute a random intensity forest
using \code{Ntree} independent spatial Intensity tree
computed with the function \code{\link{tesstree}}.

The arguments \code{X}, \code{gamma}, \code{dimyx},
and \code{test.connected} are passed to \code{\link{tesstree}}.
}
\examples{
Z <- tessforest(
  X = bei,
  gamma = 100,
  dimyx = c(101, 201),
  test.connected = FALSE,
  Ntree = 5
)
plot(Z)
}
