% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetess.R
\name{tesstree}
\alias{tesstree}
\title{Intensity tesselation tree}
\usage{
tesstree(X, lambda, target.points, test.connected = TRUE)
}
\arguments{
\item{X}{ppp}

\item{lambda}{intensity}

\item{target.points}{point a which we estimated the intensity}

\item{test.connected}{do we test if the voronoi cells are connected}
}
\value{
estimated intensity
}
\description{
Intensity tesselation tree
}
\examples{
tesstree(rpoispp(10), lambda = 50, target.points = rpoispp(10))
}
