% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesstree.R
\name{tesstree}
\alias{tesstree}
\title{Spatial Intensity tree function without covariate}
\usage{
tesstree(X, lambda = 100, dimyx = c(128, 128), test.connected = FALSE)
}
\arguments{
\item{X}{The observed data point pattern,
as a \code{\link[spatstat.geom]{ppp}} object .}

\item{lambda}{An integer. The number of points used for random tessellation.}

\item{dimyx}{A vector of two integers. The dimensions of the output image passed
to \code{\link[spatstat.geom]{as.im}}.}

\item{test.connected}{Logical. If \code{TRUE},
\code{\link[spatstat.geom]{connected}} is applied to the tessellation to split tiles
in different connected components. It is only useful if the windows of
observation od \code{X} is not convex.}
}
\value{
A pixel image, object of class \code{\link[spatstat.geom]{im.object}}.
}
\description{
Spatial Intensity tree function without covariate
}
\examples{
Z <- tesstree2(
  X = bei,
  lambda = 100,
  dimyx = c(101, 201),
  test.connected = FALSE
)
plot(Z)
}
