% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetess.R
\name{tesstree_precision}
\alias{tesstree_precision}
\title{Intensity tesselation tree high definition}
\usage{
tesstree_precision(X, lambda, target.points, test.connected = TRUE)
}
\arguments{
\item{X}{}

\item{lambda}{}

\item{target.points}{}

\item{test.connected}{}
}
\description{
Intensity tesselation tree high definition
}
