% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spforest.R
\name{vipplot}
\alias{vipplot}
\title{Vip barplot of forest}
\usage{
vipplot(x, sorted = FALSE, cores = 1, ...)
}
\arguments{
\item{x}{An spforest object}

\item{sorted}{Logical. If TRUE, the variables are sorted by importance.}

\item{cores}{To compute faster}

\item{...}{ignoted}
}
\value{
Variable importance plot
}
\description{
Vip barplot of forest
}
\examples{
forest <- RforestPP(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
vipplot(forest, sorted = TRUE)
}
