% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{vipplot}
\alias{vipplot}
\title{Vip barplot of forest}
\usage{
vipplot(x, sorted = FALSE, viptype = 4, ...)
}
\arguments{
\item{x}{A \code{\link{spforest.object}} with \code{listcov} not \code{NULL}.}

\item{sorted}{Logical. If TRUE, the variables are sorted by importance.}

\item{viptype}{An integer in \{1,2,3,4\} passed to \code{\link[spforest]{importance}}.}

\item{...}{Arguments passed to \code{\link[graphics]{barplot}}.}
}
\value{
Variable importance plot
}
\description{
Vip barplot of forest
}
\details{
First the function \code{\link[spforest]{importance}} is called to
compute the importance of each covariate in each tree \code{x}.
Then, a barplot is drawn to visualize the distribution
of the importance of each covariate, averaged over all the trees.
}
\examples{
forest <- spforest(
  X = spatstat.data::bei,
  listcovariates = spatstat.data::bei.extra,
  Ntree = 3,
  minpts = 300,
  mtry = 1
)
vipplot(forest, sorted = TRUE)
}
